+++ docs/source/conf.py
@@ -42,6 +42,7 @@ extensions = ['sphinx.ext.autodoc', 'sphinx.ext.intersphinx',
               'sphinx.ext.viewcode', 'sphinxcontrib.napoleon',
               'IPython.sphinxext.ipython_console_highlighting',
               'IPython.sphinxext.ipython_directive',
+              'sphinxcontrib.programoutput',
               ]
 
 ipython_mplbackend= None