+++ piecash/scripts/cli.py
@@ -1,0 +1,8 @@
+import click
+
+CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
+
+
+@click.group(context_settings=CONTEXT_SETTINGS)
+def cli():
+    pass