+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ In development
 ~~~~~~~~~~~~~~
 
 - add piecash CLI (refactor of scripts)
+- add book.invoices to retrieve all invoices in a book
+- expose gnucash rationals as decimals in Entry and Invoice
 
 Version 0.14.1 (2018-02-01)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~