+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ In development
 - add piecash CLI (refactor of scripts)
 - add book.invoices to retrieve all invoices in a book
 - expose gnucash rationals as decimals in Entry and Invoice
+- fix issue #65 about "template" (scheduled transactions) appearing in ledger export
+- fix issue #64 about escaping in double quote mnemonic with non alpha characters
+
 
 Version 0.14.1 (2018-02-01)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~