+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ In development
 - fix issue #65 about "template" (scheduled transactions) appearing in ledger export
 - fix issue #64 about escaping in double quote mnemonic with non alpha characters
 - fix issue #19 allowing to pass the check_same_thread flag for sqlite
+- add argument recurse=True to get_balance (fix #73)
 
 
 Version 0.14.1 (2018-02-01)